+++ darts/utils/data/shifted_dataset.py
@@ -31,7 +31,7 @@ class PastCovariatesShiftedDataset(PastCovariatesTrainingDataset):
         max_samples_per_ts: Optional[int] = None,
     ):
         """
-        A time series dataset containing tuples of (past_target, past_covariates, future_target)
+        A time series dataset containing tuples of (past_target, past_covariates, static_covariates, future_target)
         arrays, which all have length `length`.
         The "future_target" is the "past_target" target shifted by `shift` time steps forward.
         So if an emitted "past_target" (and "past_covariates") goes from position `i` to `i+length`,
@@ -83,7 +83,9 @@ class PastCovariatesShiftedDataset(PastCovariatesTrainingDataset):
     def __len__(self):
         return len(self.ds)
 
-    def __getitem__(self, idx) -> Tuple[np.ndarray, Optional[np.ndarray], np.ndarray]:
+    def __getitem__(
+        self, idx
+    ) -> Tuple[np.ndarray, Optional[np.ndarray], Optional[np.ndarray], np.ndarray]:
         return self.ds[idx]
 
 
@@ -97,7 +99,7 @@ class FutureCovariatesShiftedDataset(FutureCovariatesTrainingDataset):
         max_samples_per_ts: Optional[int] = None,
     ):
         """
-        A time series dataset containing tuples of (past_target, future_covariates, future_target)
+        A time series dataset containing tuples of (past_target, future_covariates, static_covariates, future_target)
         arrays, which all have length `length`.
         The "future_target" is the "past_target" target shifted by `shift` time steps forward.
         So if an emitted "past_target" goes from position `i` to `i+length`,
@@ -152,7 +154,9 @@ class FutureCovariatesShiftedDataset(FutureCovariatesTrainingDataset):
     def __len__(self):
         return len(self.ds)
 
-    def __getitem__(self, idx) -> Tuple[np.ndarray, Optional[np.ndarray], np.ndarray]:
+    def __getitem__(
+        self, idx
+    ) -> Tuple[np.ndarray, Optional[np.ndarray], Optional[np.ndarray], np.ndarray]:
         return self.ds[idx]
 
 
@@ -167,7 +171,7 @@ class DualCovariatesShiftedDataset(DualCovariatesTrainingDataset):
     ):
         """
         A time series dataset containing tuples of
-        (past_target, historic_future_covariates, future_covariates, future_target)
+        (past_target, historic_future_covariates, future_covariates, static_covariates, future_target)
         arrays, which all have length `length`.
         The "future_target" is the "past_target" target shifted by `shift` time steps forward.
         So if an emitted "past_target" goes from position `i` to `i+length`,
@@ -238,10 +242,22 @@ class DualCovariatesShiftedDataset(DualCovariatesTrainingDataset):
 
     def __getitem__(
         self, idx
-    ) -> Tuple[np.ndarray, Optional[np.ndarray], Optional[np.ndarray], np.ndarray]:
-        past_target, past_covariate, future_target = self.ds_past[idx]
-        _, future_covariate, _ = self.ds_future[idx]
-        return past_target, past_covariate, future_covariate, future_target
+    ) -> Tuple[
+        np.ndarray,
+        Optional[np.ndarray],
+        Optional[np.ndarray],
+        Optional[np.ndarray],
+        np.ndarray,
+    ]:
+        past_target, past_covariate, static_covariate, future_target = self.ds_past[idx]
+        _, future_covariate, _, _ = self.ds_future[idx]
+        return (
+            past_target,
+            past_covariate,
+            future_covariate,
+            static_covariate,
+            future_target,
+        )
 
 
 class MixedCovariatesShiftedDataset(MixedCovariatesTrainingDataset):
@@ -256,8 +272,7 @@ class MixedCovariatesShiftedDataset(MixedCovariatesTrainingDataset):
     ):
         """
         A time series dataset containing tuples of (past_target, past_covariates, historic_future_covariates,
-                                                    future_covariates, future_target)
-        arrays, which all have length `length`.
+        future_covariates, static_covariates, future_target) arrays, which all have length `length`.
         The "future_target" is the "past_target" target shifted by `shift` time steps forward.
         So if an emitted "past_target" goes from position `i` to `i+length`,
         the emitted "future_target" will go from position `i+shift` to `i+shift+length`.
@@ -330,16 +345,18 @@ class MixedCovariatesShiftedDataset(MixedCovariatesTrainingDataset):
         Optional[np.ndarray],
         Optional[np.ndarray],
         Optional[np.ndarray],
+        Optional[np.ndarray],
         np.ndarray,
     ]:
 
-        past_target, past_covariate, future_target = self.ds_past[idx]
-        _, historic_future_covariate, future_covariate, _ = self.ds_dual[idx]
+        past_target, past_covariate, static_covariate, future_target = self.ds_past[idx]
+        _, historic_future_covariate, future_covariate, _, _ = self.ds_dual[idx]
         return (
             past_target,
             past_covariate,
             historic_future_covariate,
             future_covariate,
+            static_covariate,
             future_target,
         )
 
@@ -355,8 +372,8 @@ class SplitCovariatesShiftedDataset(SplitCovariatesTrainingDataset):
         max_samples_per_ts: Optional[int] = None,
     ):
         """
-        A time series dataset containing tuples of (past_target, past_covariates, future_covariates, future_target)
-        arrays, which all have length `length`.
+        A time series dataset containing tuples of (past_target, past_covariates, future_covariates, static_covariates,
+        future_target) arrays, which all have length `length`.
         The "future_target" is the "past_target" target shifted by `shift` time steps forward.
         So if an emitted "past_target" goes from position `i` to `i+length`,
         the emitted "future_target" will go from position `i+shift` to `i+shift+length`.
@@ -428,10 +445,22 @@ class SplitCovariatesShiftedDataset(SplitCovariatesTrainingDataset):
 
     def __getitem__(
         self, idx
-    ) -> Tuple[np.ndarray, Optional[np.ndarray], Optional[np.ndarray], np.ndarray]:
-        past_target, past_covariate, future_target = self.ds_past[idx]
-        _, future_covariate, _ = self.ds_future[idx]
-        return past_target, past_covariate, future_covariate, future_target
+    ) -> Tuple[
+        np.ndarray,
+        Optional[np.ndarray],
+        Optional[np.ndarray],
+        Optional[np.ndarray],
+        np.ndarray,
+    ]:
+        past_target, past_covariate, static_covariate, future_target = self.ds_past[idx]
+        _, future_covariate, _, _ = self.ds_future[idx]
+        return (
+            past_target,
+            past_covariate,
+            future_covariate,
+            static_covariate,
+            future_target,
+        )
 
 
 class GenericShiftedDataset(TrainingDataset):
@@ -447,8 +476,8 @@ class GenericShiftedDataset(TrainingDataset):
         covariate_type: CovariateType = CovariateType.NONE,
     ):
         """
-        Contains (past_target, <X>_covariate, future_target), where "<X>" is past if `shift_covariates = False`
-        and future otherwise.
+        Contains (past_target, <X>_covariates, static_covariates, future_target), where "<X>" is past if
+        `shift_covariates = False` and future otherwise.
         The past chunks have length `input_chunk_length` and the future chunks have length `output_chunk_length`.
         The future chunks start `shift` after the past chunks' start.
 
@@ -518,11 +547,13 @@ class GenericShiftedDataset(TrainingDataset):
     def __len__(self):
         return self.ideal_nr_samples
 
-    def __getitem__(self, idx) -> Tuple[np.ndarray, Optional[np.ndarray], np.ndarray]:
+    def __getitem__(
+        self, idx
+    ) -> Tuple[np.ndarray, Optional[np.ndarray], np.ndarray, Optional[np.ndarray]]:
         # determine the index of the time series.
-        ts_idx = idx // self.max_samples_per_ts
-        ts_target = self.target_series[ts_idx]
-        target_vals = ts_target.random_component_values(copy=False)
+        target_idx = idx // self.max_samples_per_ts
+        target_series = self.target_series[target_idx]
+        target_vals = target_series.random_component_values(copy=False)
 
         # determine the actual number of possible samples in this time series
         n_samples_in_ts = len(target_vals) - self.size_of_both_chunks + 1
@@ -531,22 +562,24 @@ class GenericShiftedDataset(TrainingDataset):
             n_samples_in_ts >= 1,
             "The dataset contains some time series that are too short to contain "
             "`max(self.input_chunk_length, self.shift + self.output_chunk_length)` "
-            "({}-th series)".format(ts_idx),
+            "({}-th series)".format(target_idx),
         )
 
         # determine the index at the end of the output chunk
         # it is originally in [0, self.max_samples_per_ts), so we use a modulo to have it in [0, n_samples_in_ts)
         end_of_output_idx = (
-            len(ts_target)
-            - (idx - (ts_idx * self.max_samples_per_ts)) % n_samples_in_ts
+            len(target_series)
+            - (idx - (target_idx * self.max_samples_per_ts)) % n_samples_in_ts
         )
 
         # optionally, load covariates
-        ts_covariate = self.covariates[ts_idx] if self.covariates is not None else None
+        covariate_series = (
+            self.covariates[target_idx] if self.covariates is not None else None
+        )
 
-        main_cov_type = CovariateType.NONE
+        main_covariate_type = CovariateType.NONE
         if self.covariates is not None:
-            main_cov_type = (
+            main_covariate_type = (
                 CovariateType.FUTURE if self.shift_covariates else CovariateType.PAST
             )
 
@@ -556,17 +589,17 @@ class GenericShiftedDataset(TrainingDataset):
             past_end,
             future_start,
             future_end,
-            cov_start,
-            cov_end,
+            covariate_start,
+            covariate_end,
         ) = self._memory_indexer(
-            ts_idx=ts_idx,
-            ts_target=ts_target,
+            target_idx=target_idx,
+            target_series=target_series,
             shift=self.shift,
             input_chunk_length=self.input_chunk_length,
             output_chunk_length=self.output_chunk_length,
             end_of_output_idx=end_of_output_idx,
-            ts_covariate=ts_covariate,
-            cov_type=main_cov_type,
+            covariate_series=covariate_series,
+            covariate_type=main_covariate_type,
         )
 
         # extract sample target
@@ -577,13 +610,13 @@ class GenericShiftedDataset(TrainingDataset):
         covariate = None
         if self.covariates is not None:
             raise_if_not(
-                cov_end <= len(ts_covariate),
-                f"The dataset contains {main_cov_type.value} covariates "
+                covariate_end <= len(covariate_series),
+                f"The dataset contains {main_covariate_type.value} covariates "
                 f"that don't extend far enough into the future. ({idx}-th sample)",
             )
 
-            covariate = ts_covariate.random_component_values(copy=False)[
-                cov_start:cov_end
+            covariate = covariate_series.random_component_values(copy=False)[
+                covariate_start:covariate_end
             ]
 
             raise_if_not(
@@ -593,9 +626,10 @@ class GenericShiftedDataset(TrainingDataset):
                     if self.shift_covariates
                     else self.input_chunk_length
                 ),
-                f"The dataset contains {main_cov_type.value} covariates "
+                f"The dataset contains {main_covariate_type.value} covariates "
                 f"whose time axis doesn't allow to obtain the input (or output) chunk relative to the "
                 f"target series.",
             )
 
-        return past_target, covariate, future_target
+        static_covariate = target_series.static_covariates_values(copy=False)
+        return past_target, covariate, static_covariate, future_target