+++ CHANGELOG.md
@@ -11,6 +11,11 @@ Darts is still in an early development phase and we cannot always guarantee back
 - Option to avoid global matplotlib configuration changes.
   [#924](https://github.com/unit8co/darts/pull/924) by [Mike Richman](https://github.com/zgana).
 - Model Improvements: Option for changing activation function for NHiTs and NBEATS. NBEATS support for dropout. NHiTs Support for AvgPooling1d. [#955](https://github.com/unit8co/darts/pull/955) by [Greg DeVos](https://github.com/gdevos010)
+- New datasets: ETT and Electricity. [#617](https://github.com/unit8co/darts/pull/617)
+  by [Greg DeVos](https://github.com/gdevos010)
+- Implemented ["GLU Variants Improve Transformer"](https://arxiv.org/abs/2002.05202) for transformer based models (transformer and TFT). [#959](https://github.com/unit8co/darts/issues/959)
+  by [Greg DeVos](https://github.com/gdevos010) 
+- Added support for torch metrics during training and validation. [#996](https://github.com/unit8co/darts/pull/996) by [Greg DeVos](https://github.com/gdevos010)
 
 ## [0.19.0](https://github.com/unit8co/darts/tree/0.19.0) (2022-04-13)
 ### For users of the library:
@@ -32,6 +37,7 @@ Darts is still in an early development phase and we cannot always guarantee back
 - New notebook showcasing transfer learning and training forecasting models on large time
   series datasets. [#885](https://github.com/unit8co/darts/pull/885) 
   by [Julien Herzen](https://github.com/hrzn).
+
 
 **Fixed**
 - Some issues with PyTorch Lightning >= 1.6.0 [#888](https://github.com/unit8co/darts/pull/888)